Fix bug in blox connector while executing block
The `_run_command()` method was a generator. Hence, it wasn't being
called when no output was produced. To fix it, we introduce a separate
method `_process_lines()`, when output is important (returned lazily as
before via generator). The method `_run_command()` is now used only when
no output is expected.